+++ Changelog.md
@@ -27,7 +27,7 @@
 * Parsing of Columns spacing
 * Redone parsing `gridSpan` and `vMerge`
 * Add ability to get style on which style is based
-* Parsing `TableStyleColumnBandSize` and `TableStyleRowBandSize`
+* Parsing `TableStyleColumnBandSize`, `TableStyleRowBandSize`, `TableLayout`
 
 ### Fixes
 * Fix parsing document style id - it can be string, not only digit